Currently have setup with _app.js holding the cookie related functions
