+++ src/pages/_app.js
@@ -3,9 +3,22 @@ import 'bootstrap/dist/css/bootstrap.css';
 import Head from 'next/head';
 import Script from 'next/script';
 import Layout from '../../components/Layout.js';
+import react, {useState, useEffect} from 'react';
+import Cookies from './Login/setCookie.js';
 
 
 export default function App({ Component, pageProps }) {
+  let [cookie, setCookie] = useState('');
+
+  useEffect(() => {
+    updateCookie();
+    console.log('herllo');
+    console.log('Cookie has changed:', cookie)
+  }, [cookie]);
+
+  const updateCookie = () => {
+    setCookie(Cookies.getCookie());
+  }
   return (
     <>
       <Head>
@@ -13,7 +26,7 @@ export default function App({ Component, pageProps }) {
       </Head>
 
       <Layout>
-        <Component {...pageProps} />
+        <Component {...pageProps} cookie={cookie} setCookie={setCookie} updateCookie={updateCookie.bind(this)}/>
       </Layout>
